fix(startpage)[OK-39] Remove duplicated code, update font, rename to startpage.module.css
