+++ frontend/src/components/startpage/startpage.module.css
@@ -9,7 +9,7 @@
     position: absolute;
     left: 0;
     top: 50%;
-    transform: translateY(-65%);
+    transform: translateY(-50%);
     width: 5px; 
     height: 100%;
     background-color: #F2E2BC;
